requirements.txt
django
openai
google-genai
Pillow
python-dotenv
